Palm Rust: Change obsidian break time from 4.2s to 5.8s
Over the years, this map has usually been rushed very quickly - by the time someone finishes crafting armor, the obsidian can already broken due to a 4 second break-time (3.5 seconds with enabled beacon). Majority of matches on this map are sub-10 minutes, averaging somewhere around 6.

I'd like to give it a shot to up the time to 5.8s by downgrading the efficiency - if this ends up backfiring, we can change it back. But measuring the blocks to distance in singleplayer, this change should be fine.